refactor: extract terminal-cause machinery to terminal.rs (Phase 2)
Move Group C (ConnTerminalSlot/State, record/observe/commit_conn, SendTerminalSlot,
load_winner, publish_send, classify_transport/conn_shutdown, peek_conn_terminal)
and the connection_terminal test module out of lib.rs into terminal.rs, re-exported
at the crate root (new_conn_terminal_slot re-export cfg(test)-gated). Private fns +
ConnTerminalState::{observe,has_internal} raised to pub(crate). Pure move; no
behavior/API change. 176 tests / 2 ignored.
